move output classes to separate file
