+++ pysd/py_backend/model.py
@@ -8,14 +8,14 @@ a separate file).
 import warnings
 import inspect
 import pickle
-import time as t
 from pathlib import Path
 from typing import Union
 
 import numpy as np
 import xarray as xr
 import pandas as pd
-import netCDF4 as nc
+
+from pysd._version import __version__
 
 from . import utils
 from .statefuls import DynamicStateful, Stateful
@@ -24,8 +24,7 @@ from .cache import Cache, constant_cache
 from .data import TabData
 from .lookups import HardcodedLookups
 from .components import Components, Time
-
-from pysd._version import __version__
+from .output import ModelOutput
 
 
 class Macro(DynamicStateful):
@@ -1025,6 +1024,7 @@ class Model(Macro):
         self.time.set_control_vars(**self.components._control_vars)
         self.data_files = data_files
         self.missing_values = missing_values
+        self.progress = None
         if initialize:
             self.initialize()
 
@@ -1673,7 +1673,6 @@ class Model(Macro):
         """
         self.state = self.state + self.ddt() * dt
 
-
     def _integrate(self, out_obj):
         """
         Performs euler integration and writes results to the out_obj.
@@ -1713,326 +1712,3 @@ class Model(Macro):
             out_obj.update(self)
 
         progressbar.finish()
-
-
-class OutputHandler():
-    """
-    Interface for the different output handlers.
-    """
-
-    def initialize(self, model, capture_elements):
-        pass
-
-    def update(self, model, capture_elements):
-        pass
-
-    def postprocess(self, **kwargs):
-        pass
-
-    def add_run_elements(self, model, capture_elements):
-        pass
-
-
-class DatasetHandler(OutputHandler):
-    """
-    Manages simulation results stored as netCDF4 Dataset.
-    """
-
-    def __init__(self, out_file):
-        self.out_file = out_file
-        self.step = 0
-        self.ds = None
-
-    def initialize(self, model, capture_elements):
-        """
-        Creates a netCDF4 Dataset and adds model dimensions and variables
-        present in the capture elements to it.
-
-        Parameters
-        ----------
-        model: pysd.Model
-            PySD Model object
-        capture_elements: set
-            Which model elements to capture - uses pysafe names.
-
-        Returns
-        -------
-        None
-
-        """
-        self.ds = nc.Dataset(self.out_file, "w")
-        # defining global attributes
-        self.ds.description = "Results for simulation run on" \
-        f"{t.ctime(t.time())} using PySD version {__version__}"
-        self.ds.model_file = model.py_model_file
-        self.ds.timestep = "{}".format(model.components.time_step())
-        self.ds.initial_time = "{}".format(model.components.initial_time())
-        self.ds.final_time = "{}".format(model.components.final_time())
-
-        # creating variables for all model dimensions
-        for dim_name,coords in model.subscripts.items():
-            coords = np.array(coords)
-            # create dimension
-            self.ds.createDimension(dim_name, len(coords))
-            # length of the longest string in the coords
-            max_str_len = len(max(coords, key=len))
-
-            # create variable
-            # TODO: check if the type could be defined otherwise)
-            var = self.ds.createVariable(dim_name,
-                                    f"S{max_str_len}",
-                                    (dim_name,))
-            # assigning values to variable
-            var[:] = coords
-
-        # creating the time dimension as unlimited
-        self.ds.createDimension("time", None)
-
-        # creating variables in capture_elements
-        self.__create_ds_vars(model, capture_elements)
-
-    def update(self, model, capture_elements):
-        """
-        Writes values of variables in capture_elements in the netCDF4 Dataset.
-
-        Parameters
-        ----------
-        model: pysd.Model
-            PySD Model object
-        capture_elements: set
-            Which model elements to capture - uses pysafe names.
-
-        Returns
-        -------
-        None
-        """
-        for key in capture_elements:
-            comp = getattr(model.components, key)
-            comp_vals = comp()
-            if isinstance(comp_vals, xr.DataArray):
-                self.ds[key][self.step, :] = comp_vals.values
-            elif isinstance(comp_vals, np.ndarray):
-                self.ds[key][self.step, :] = comp_vals
-            else:
-                self.ds[key][self.step] = comp_vals
-
-        self.step += 1
-
-    def postprocess(self, **kwargs):
-        """
-        Closes netCDF4 Dataset.
-        """
-        self.ds.close()
-
-        if kwargs.get("flatten"):
-            warnings.warn("DataArrays stored in netCDF4 will not be flattened")
-
-        print(f"Results stored in {self.out_file}")
-
-    def add_run_elements(self, model, capture_elements):
-        """
-        Adds constant elements to netCDF4 Dataset.
-        Parameters
-        ----------
-        model: pysd.Model
-            PySD Model object
-        capture_elements: list
-            List of constant elements
-        Returns
-        -------
-        None
-        """
-        # creating variables in capture_elements
-        # TODO we are looping through all capture elements twice. This
-        # could be avoided
-        self.__create_ds_vars(model, capture_elements)
-
-        for key in capture_elements:
-            comp = getattr(model.components, key)
-            comp_vals = comp()
-            try:
-                for num,_ in enumerate(self.ds["time"][:]):
-                    if isinstance(comp_vals, xr.DataArray):
-                        self.ds[key][num, :] = comp_vals.values
-                    elif isinstance(comp_vals, np.ndarray):
-                        if comp_vals.size == 1:
-                            self.ds[key][num] = comp_vals
-                        else:
-                            self.ds[key][num, :] = comp_vals
-                    else:
-                        self.ds[key][num] = comp_vals
-            except ValueError:
-                warnings.warn(f"The dimensions of {key} in the results "
-                              "do not match the declared dimensions for this "
-                              "variable. The resulting values will not be "
-                              "included in the results file.")
-
-    def __create_ds_vars(self, model, capture_elements):
-        """
-        Create new variables in a netCDF4 Dataset from the capture_elements.
-
-        Parameters
-        ----------
-        model: pysd.Model
-            PySD Model object
-        capture_elements: set
-            Which model elements to capture - uses pysafe names.
-
-        Returns
-        -------
-        None
-
-        """
-
-        for key in capture_elements:
-            comp = getattr(model.components, key)
-            comp_vals = comp()
-
-            if isinstance(comp_vals, (xr.DataArray, np.ndarray)):
-                if comp.subscripts:
-                    dims = ("time",) + tuple(comp.subscripts)
-                else:
-                    dims = ("time",)
-            else:
-                dims = ("time",)
-
-            self.ds.createVariable(key, "f8", dims)
-
-            # adding units and description as metadata for each var
-            self.ds[key].units = model.doc.loc[
-                        model.doc["Py Name"] == key,
-                        "Units"].values[0] or "Missing"
-            self.ds[key].description = model.doc.loc[
-                        model.doc["Py Name"] == key,
-                        "Comment"].values[0] or "Missing"
-
-
-class DataFrameHandler(OutputHandler):
-    """
-    Manages simulation results stored as pandas DataFrame.
-    """
-    def __init__(self):
-        self.ds = None
-
-    def initialize(self, model, capture_elements):
-        """
-        Creates a pandas DataFrame and adds model variables as columns
-
-        Parameters
-        ----------
-        model: pysd.Model
-            PySD Model object
-        capture_elements: set
-            Which model elements to capture - uses pysafe names.
-
-        Returns
-        -------
-        None
-
-        """
-        self.ds =  pd.DataFrame(columns=capture_elements)
-
-        return None
-
-    def update(self, model, capture_elements):
-        """
-        Add a row to the results pandas DataFrame with the values of the
-        variables listed in capture_elements.
-
-        Parameters
-        ----------
-        model: pysd.Model
-            PySD Model object
-        capture_elements: set
-            Which model elements to capture - uses pysafe names.
-
-        Returns
-        -------
-        None
-        """
-
-        self.ds.at[model.time.round()] = [
-            getattr(model.components, key)()
-            for key in capture_elements]
-
-    def postprocess(self, **kwargs):
-        """
-        Delete time column from the pandas DataFrame and flatten xarrays if
-        required.
-
-        Returns
-        -------
-        ds: pandas.DataFrame
-            Simulation results stored as a pandas DataFrame.
-        """
-        # delete time column as it was created only for avoiding errors
-        # of appending data. See previous TODO.
-        del self.ds["time"]
-
-        return utils.make_flat_df(self.ds,
-                                  kwargs["return_addresses"],
-                                  kwargs["flatten"])
-
-    def add_run_elements(self, model, capture_elements):
-        """
-        Adds constant elements to a dataframe.
-
-        Parameters
-        ----------
-        model: pysd.Model
-            PySD Model object
-        capture_elements: set
-            Which model elements to capture - uses pysafe names.
-
-        Returns
-        -------
-        None
-        """
-        nx = len(self.ds.index)
-        for element in capture_elements:
-            self.ds[element] = [getattr(model.components, element)()] * nx
-
-
-class ModelOutput():
-    """
-    Handles different types of outputs by dispatchinging the tasks to adequate
-    object handlers.
-
-    Parameters
-    ----------
-    model: pysd.Model
-        PySD Model object
-    capture_elements: set
-        Which model elements to capture - uses pysafe names.
-    out_file: str or pathlib.Path
-        Path to the file where the results will be written.
-    """
-    valid_output_files = [".nc"]
-
-    def __init__(self, model, capture_elements, out_file=None):
-
-        if out_file:
-            if out_file.suffix == ".nc":
-                self.handler = DatasetHandler(out_file)
-            else:
-                raise ValueError(
-                    f"Unsupported output file format {out_file.suffix}")
-        else:
-            self.handler = DataFrameHandler()
-
-        capture_elements.add("time")
-        self.capture_elements = capture_elements
-
-        self.initialize(model)
-
-    def initialize(self, model):
-        self.handler.initialize(model, self.capture_elements)
-
-    def update(self, model):
-        self.handler.update(model, self.capture_elements)
-
-    def postprocess(self, **kwargs):
-        return self.handler.postprocess(**kwargs)
-
-    def add_run_elements(self, model, run_elements):
-        self.handler.add_run_elements(model, run_elements)
